+++ frontend/src/vite-env.d.ts
@@ -1,5 +1,5 @@
 /// <reference types="vite/client" />
-declare module "*.css" {
+declare module '*.css' {
   const classes: { [key: string]: string };
   export default classes;
 }